requirements.txt
certifi==2019.3.9
chardet==3.0.4
config==0.4.2
gevent==1.4.0
greenlet==0.4.15
idna==2.7
paho-mqtt==1.4.0
PyBluez==0.22
requests==2.20.0
RPi.GPIO==0.6.5
six==1.12.0
sonoff-python==0.1
urllib3==1.24.1
websocket==0.2.1
websocket-client==0.56.0
